Update on ticket #2, This should now work for simple one off entries on the scheduler in terms of blocking in EVE time or local time.
Still have to work out how to code for the more complex repeat schedule entries.

Also I've left the hardcoded DT entry in for now.

Anyone know what the 'Silent' option is for?

Oh, and I also fixed a bug brough in by patch [423] as the label showing the collision wasn't added to a panel.