OK, here's a load of notes on what gives with my skill completion coding.
Also, I'm fed up of seeing GroupId in some places and not others resulting in it not being carried through properly.

With the soon to be written implementation of uploading to inEVE and such, I figured having the groupId's might not be a bad idea, so they are now used correctly.

They have also been added to the skills.xml.gz file and associated loading code.