A fresh upload of my current implant group code, minus the menu option itself to open the screens.
This will help to test out some of the transition code updating people from the old implant details to the new ones.

btw, some of this code is simply so the old system still works while we transfer to the new system.

If folks could please test this by having a look at their sttings.xml file, I'd appreciate it.