And behold, for the man did deliver.
We now have implant sets. Please for the love of the cute furry hamsters, please someone sanity check this lot.

I made a few adjustments so that we didn't have the same code cropping up half a dozen times for things like working out the current attributes or working out which slot was which attribute and vice-versa. Looking at the code I got the impression that there might be a few places where we can do that sort of thing.