This patch adds prices to skills, and the ability to mark skills as owned.
The Skills.xml has been modified to include cost - the prices are from the datadump so may not be 100% accurate but it gives a good idea.

You can mark a skill as owned either in the skill browser, or by  the context menu or the plan viewer.

The plan window shows the total cost of all unnown and unowned skills for the plan in the status bar.

The copy to clipboard (and IGB) has been modified to optionally include the cost of unknown & unowned skills, and a total cost.

Known skills are saved in the settings.xml file.