Fixes a fluffy discovered by Eewec - when we switch plans in the NewPLannerWindow control, we're not setting the plan's reference to the NewPlannerWindow.
This also fixes an as ye unreported bug - if you switch plans within the NewPlannerWindow, and then delete the switched to plan from the same control, the planner window won't close.

While I was here, I also noticed that we didn't close the implant calc when closing the planner window. That's fixed too.