The default portrait folder is currently set to:  * Local Application Data\CCP\EVE\cache\Pictures\Portraits
However it is now (under Vista 64):
 * Local Application Data\CCP\EVE\c_program_files_(x86)_ccp_eve_tranquility\cache\Pictures\Portraits

Changes make the process of getting the portrait from the EVE Cache more robust. EVEMon will now try and locate the CCP\EVE\cache directory by looking for a directory that ends in "_tranquility". Adapted the "Set Portrait Folder" dialog to disable the default option in the event that a default folder can not be located.